python-backend/requirements.txt
openai-agents
python-dotenv
pydantic
py-mini-racer
supabase>=2.0.3
openai>=1.6.1
psycopg2-binary>=2.9.9
pyyaml>=6.0
